refactor(security): extract URL redaction
